+++ src/stateAggregator/stateAggregator.ts
@@ -6,25 +6,29 @@
  */
 
 import { AsyncOptionalCreatable } from '@salesforce/kit';
-import { Optional } from '@salesforce/ts-types';
+import { Global } from '../global';
 import { AliasAccessor } from './accessors/aliasAccessor';
 import { OrgAccessor } from './accessors/orgAccessor';
 import { SandboxAccessor } from './accessors/sandboxAccessor';
 import { TokenAccessor } from './accessors/tokenAccessor';
-
 export class StateAggregator extends AsyncOptionalCreatable {
-  private static instance: Optional<StateAggregator>;
-
+  private static instanceMap: Map<string, StateAggregator> = new Map();
   public aliases!: AliasAccessor;
   public orgs!: OrgAccessor;
   public sandboxes!: SandboxAccessor;
   public tokens!: TokenAccessor;
 
+  /**
+   * Reuse a StateAggregator if one was already created for the current global state directory
+   * Otherwise, create one and adds it to map for future reuse.
+   * HomeDir might be stubbed in tests
+   */
   public static async getInstance(): Promise<StateAggregator> {
-    if (!StateAggregator.instance) {
-      StateAggregator.instance = await StateAggregator.create();
+    if (!StateAggregator.instanceMap.has(Global.DIR)) {
+      StateAggregator.instanceMap.set(Global.DIR, await StateAggregator.create());
     }
-    return StateAggregator.instance;
+    // TS assertion is valid because there either was one OR it was just now instantiated
+    return StateAggregator.instanceMap.get(Global.DIR) as StateAggregator;
   }
 
   /**
@@ -32,8 +36,8 @@ export class StateAggregator extends AsyncOptionalCreatable {
    *
    * *NOTE: Only call this method if you must and you know what you are doing.*
    */
-  public static clearInstance(): void {
-    delete StateAggregator.instance;
+  public static clearInstance(path = Global.DIR): void {
+    StateAggregator.instanceMap.delete(path);
   }
 
   protected async init(): Promise<void> {